Import first prototype of GeoStyler GeoJSON parser
This imports a first prototype of a GeoJSON parser for the GeoStyler
project (https://github.com/terrestris/geostyler).
The parser reads a given GeoJSON FeatureCollection and transforms it
into the internal data structure of the GeoStyler software.

At the moment the Data Model classes of GeoStyler a placed directly in
this repository. By a follow up commit they will be used as dependency
once the correspondig package is available.