Altern. approach f. interface for GeoStyler parser
Implements a different approach for the design of the parser interface
for GeoStyler parser.
This models the Data Model as interface and adds the 'readData' method to
tis interface. So no further interface class is needed. In consequence
the Data object fills itself in the concrete implementation, when the
'readData' method is executed.